Authorization added

+++ blogpost/backend/routes/posts.js
@@ -34,7 +34,8 @@ router.post("", checkAuth ,multer({storage}).single("image") ,(req, res, next)=>
 
   const post = new Post({title: req.body.title,
                         content: req.body.content,
-                        imagePath: url + "/images/" + req.file.filename
+                        imagePath: url + "/images/" + req.file.filename,
+                        creator: req.userData.userId
                       });
 
   post.save().then( (createdPost) => {
@@ -73,9 +74,13 @@ router.get("", (req, res) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then( (result) => {
-    console.log(result);
-    res.status(200).json({message: "Post Deleted"});
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then( (result) => {
+
+    if(result.deletedCount > 0) {
+      res.status(200).json({message: "Delete Successful"});
+    } else {
+      res.status(401).json({message: "Not Authorized to Delete"});
+    }
   })
 });
 
@@ -87,12 +92,20 @@ router.put("/:id", checkAuth,multer({storage}).single("image"), (req, res, next)
     imagePath =  url + "/images/" + req.file.filename
   }
 
-  const post = new Post({_id:req.body.id, title: req.body.title, content: req.body.content, imagePath: imagePath});
+  const post = new Post({_id:req.body.id,
+                          title: req.body.title,
+                          content: req.body.content,
+                          imagePath: imagePath,
+                          creator: req.userData.userId});
 
-  console.log(post);
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then( (result) => {
 
-  Post.updateOne({_id: req.params.id}, post).then( (result) => {
+    if(result.matchedCount > 0) {
       res.status(200).json({message: "Update Successful"});
+    } else {
+      res.status(401).json({message: "Not Authorized to Edit"});
+    }
+
   })
 });
